+++ CET_opt_2.0_main/CET-impl/src/Traversal/BFSGraphTraversal.java
@@ -2,7 +2,7 @@ package src.Traversal;
 
 import src.Components.CompressedGraph;
 import src.util.CustomDS.ArrayQueue;
-import src.util.CustomDS.IntArray;
+import src.util.CustomDS.ShortArray;
 
 public class BFSGraphTraversal extends GraphTraversal {
     public BFSGraphTraversal(CompressedGraph graph, boolean saveToMem) {
@@ -10,10 +10,10 @@ public class BFSGraphTraversal extends GraphTraversal {
         traversalType = TraversalType.BFS;
     }
 
-    public void traversal(int start) {
-        ArrayQueue<IntArray> queue = new ArrayQueue<>(graph.getNumVertex());
+    public void traversal(short start) {
+        ArrayQueue<ShortArray> queue = new ArrayQueue<>(graph.getNumVertex());
 
-        IntArray path = new IntArray(1);
+        ShortArray path = new ShortArray(1);
         path.add(start);
 
         queue.offer(path);
@@ -24,13 +24,13 @@ public class BFSGraphTraversal extends GraphTraversal {
         }
 
         while (!queue.isEmpty()) {
-            IntArray currentPath = queue.poll();
+            ShortArray currentPath = queue.poll();
             int cur = currentPath.getLast();
             for(int i = graph.rowIndex[cur]; i < graph.rowIndex[cur + 1]; i++) {
-                int neighbour = graph.colIndex[i];
+                short neighbour = (short) graph.colIndex[i];
 
                 // FIXME: java heap space out of memory error when the computation
-                IntArray newArray = new IntArray(currentPath);
+                ShortArray newArray = new ShortArray(currentPath);
                 newArray.add(neighbour);
                 if(graph.endContains(neighbour)) {
                     if(isSaveToMem) validPaths.add(newArray.getArray());
